Played around with two Return components
Added a div within the Pause button div and changed the CSS to push it to the bottom and somewhat center it.
Added a return SVG component that hides itself when the screen height is larger than the width, and did vice versa for the return SVG component at the top.